requirements.txt
diffusers
transformers
torch
safetensors
Pillow
accelerate
streamlit
